requirements-test.txt
pip>=21.0,<24.3
colorlog==6.8.2
ruff==0.6.3
pydocstyle==6.3.0
isort==5.13.2
aiounifi==80
segno==1.6.1
Pillow==10.4.0
pytest-homeassistant-custom-component==0.13.158
